backend/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
python-dotenv
cors
